requirements.txt
fastapi
requests
python-dotenv
motor
